feat: added feasibility report
