--- src/interface/views/network/NetworkDataFeasibilityReport/index.vue
@@ -1,11 +1,0 @@
-<template>
-  <div></div>
-</template>
-
-<script>
-export default {
-  name: "Index",
-};
-</script>
-
-<style scoped></style>
